chg GetHeightFromTerrain to return a float
git-svn-id: http://vtp.googlecode.com/svn/trunk@3496 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainApps/VTBuilder/Frame.cpp
@@ -1089,9 +1089,9 @@ bool MainFrame::SampleCurrentImages(vtImageLayer *pTarget)
 }
 
 
-double MainFrame::GetHeightFromTerrain(DPoint2 &p)
+float MainFrame::GetHeightFromTerrain(DPoint2 &p)
 {
-	double height = INVALID_ELEVATION;
+	float height = INVALID_ELEVATION;
 
 	int layers = m_Layers.GetSize();
 	for (int i = 0; i < layers; i++)
